requirements.txt
Flask
Flask-CORS
psycopg2-binary
python-dotenv
pdfplumber
google-generativeai
langchain
langchain-google-genai